requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0
plexapi>=4.15.0
requests>=2.31.0 